fix: move app inside restapi - and cleanup and rename
