--- implementation/Assembler/src/main/java/module-info.java
@@ -1,9 +1,0 @@
-module assembler_module{
-
-    requires persistence_module;
-    requires businesslogic_module;
-    requires restapi_module;
-    requires quarkus.core;
-
-    
-}
