--- implementation/openapi-swaggerui-quickstart/.mvn/wrapper/.gitignore
@@ -1,1 +1,0 @@
-maven-wrapper.jar
