--- implementation/openapi-swaggerui-quickstart/src/main/java/org/acme/openapi/swaggerui/Fruit.java
@@ -1,15 +1,0 @@
-package org.acme.openapi.swaggerui;
-
-public class Fruit {
-
-    public String name;
-    public String description;
-
-    public Fruit() {
-    }
-
-    public Fruit(String name, String description) {
-        this.name = name;
-        this.description = description;
-    }
-}
