--- implementation/openapi-swaggerui-quickstart/src/main/java/org/acme/openapi/swaggerui/FruitResource.java
@@ -1,37 +1,0 @@
-package org.acme.openapi.swaggerui;
-
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.DELETE;
-import jakarta.ws.rs.Path;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.Set;
-
-@Path("/fruits")
-public class FruitResource {
-
-    private Set<Fruit> fruits = Collections.newSetFromMap(Collections.synchronizedMap(new LinkedHashMap<>()));
-
-    public FruitResource() {
-        fruits.add(new Fruit("Apple", "Winter fruit"));
-        fruits.add(new Fruit("Pineapple", "Tropical fruit"));
-    }
-
-    @GET
-    public Set<Fruit> list() {
-        return fruits;
-    }
-
-    @POST
-    public Set<Fruit> add(Fruit fruit) {
-        fruits.add(fruit);
-        return fruits;
-    }
-
-    @DELETE
-    public Set<Fruit> delete(Fruit fruit) {
-        fruits.removeIf(existingFruit -> existingFruit.name.contentEquals(fruit.name));
-        return fruits;
-    }
-}
