+++ implementation/restapi/pom.xml
@@ -1,14 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.acme</groupId>
-    <artifactId>openapi-swaggerui-quickstart</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
     <parent>
-        <artifactId>airlineinformationsystem</artifactId>
         <groupId>io.github.fontysvenlo.ais</groupId>
+        <artifactId>airlineinformationsystem</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
+    <groupId>io.github.fontysvenlo.ais.restapi</groupId>
+    <artifactId>restapi</artifactId>
     <properties>
         <compiler-plugin.version>3.13.0</compiler-plugin.version>
         <maven.compiler.release>21</maven.compiler.release>
