+++ implementation/restapi/src/test/java/org/acme/openapi/swaggerui/FruitResourceTest.java
